\name{diprom-package}
\alias{diprom-package}
\alias{diprom}
\docType{package}
\title{
\packageTitle{diprom}
}
\description{
\packageDescription{diprom}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{diprom}
\packageIndices{diprom}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{diprom}

Maintainer: \packageMaintainer{diprom}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one ~~
~~ per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
